Rename kinto submodules to remotesettings
